requirements.txt
streamlit==1.47.0             # Streamlit UI framework
pytesseract==0.3.12           # OCR engine wrapper for Tesseract
Pillow==10.3.0                # Image processing and manipulation
numpy==2.2.6                  # Numerical computing (ML + stats)
pandas==2.2.2                 # Data processing and analysis
scikit-learn==1.3.0           # ML algorithms (Logistic Regression, TF-IDF)
nltk==3.8.1                   # NLP processing (tokenizing, stopwords, etc.)
requests==2.32.3              # For API calls and web requests
tldextract==4.5.1             # Domain extraction for URL analysis
whois==0.9.27                 # WHOIS lookup for domain intelligence
python-dotenv==1.0.0          # Environment variable management
joblib==1.4.2                 # Model saving/loading
shap